+++ shardingsphere-underlying/shardingsphere-route/pom.xml
@@ -27,4 +27,12 @@
     </parent>
     <artifactId>shardingsphere-route</artifactId>
     <name>${project.artifactId}</name>
+    
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>shardingsphere-sql-parser-relation</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
 </project>
